Class Routine: Using HTML and CSS
This Project is a simple class routine created using HTML and CSS. HTML was used to structure the routine in a table format, while CSS was applied to style it for better and a clean layout. The routine clearly displays subjects, time and days along with subject teacher's name, making it easy to understand and visually organized.